--- index.html
@@ -1,68 +1,0 @@
-<html>
-    <head>
-        <script src="./SDL.min.js"></script>
-        <style>
-            .parameter {
-                display: flex;
-                flex-direction: row;
-                align-items: center;
-            }
-
-            .structParameter {
-                display: flex;
-                flex-direction: column;
-                align-items: center;
-                width: 90%;
-            }
-
-            .structParameter .parameter {
-                width: 100%;
-                margin-left: 10%;
-            }
-        </style>
-    </head>
-    <body>
-        <script type="module" src="client/main.js">
-        </script>
-
-        <div id="main" style="display: flex; flex-direction: row; height: 70%;">
-            <div id="registration" style="display: flex; flex-direction: column; flex: 1; height: 100%;">
-                <input id="regAppId" type="text" value="sdlpt-js">
-                <input id="regAppName" type="text" value="SDLPT">
-    
-                <select multiple id="regAppHmiTypes">
-                </select>
-    
-                <input id="regCoreWsURI" type="text" value="wss://a59nfp.mstaging.sdl.tools">
-                <input id="regCoreWsPort" type="text" value="444">
-    
-                <button id="regConnect">Connect App</button>
-
-                <br>
-
-                <input id="specBranch" type="text" value="develop">
-                <input id="specRemote" type="text" value="smartdevicelink">
-                <button id="updateSpec">Update API Spec</button>
-
-                <br>
-                
-                <h5 id="hmiStatus"></h5>
-            </div>
-    
-            <div id="rpc" style="display: flex; flex-direction: column; flex: 4; justify-content: center; padding: 5%;">
-                <label for="rpcs">RPC: <input id="irpcs" list="rpcs" name="rpcs" type="text"></label>
-                <datalist id="rpcs"></datalist>
-
-                <div id="rpcForm" style="overflow: scroll; height: 100%;">
-
-                </div>
-
-                <button id="sendRpc">SEND</button>
-            </div>
-        </div>
-
-        <div id="history" style="display: flex; flex-direction: column; height: 30%; overflow: scroll;">
-
-        </div>
-    </body>
-</html>
